+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 All notable changes to the "simvolio-support" extension will be documented in this file.
 
+## 0.3.0
+
++ basic simvolio complete helper
+
 ## 0.2.0
 
 + separation of languages (simvolio-protypo => simvolio, protypo)
 
 ## 0.1.0
 
-+ formatting and complete helpers
++ formatting and protypo complete helpers
 
 ## [0.0.0]